Fixed Forge entity data serializers using load-order-dependent IDs when other mods register their own serializers. Fixed hurt, flash, and shader overlays being dropped by entity, block entity, and item renderers. Fixed offset or rotated models being culled with incorrect visibility bounds. Fixed touching faces disappearing from models whose textures can contain transparent pixels. Improved compatibility with EntityCulling for entities and block entities using extended model bounds.

+++ Common/src/main/java/de/markusbordihn/easymodelentities/client/render/EasyModelAnimationVariantTracker.java
@@ -69,9 +69,8 @@ final class EasyModelAnimationVariantTracker<T> {
     } else {
       variant.rollCounter++;
       int index =
-          (int)
-              Math.floorMod(
-                  mix(roll.seed(), roll.baseName().hashCode(), variant.rollCounter), size);
+        Math.floorMod(
+            mix(roll.seed(), roll.baseName().hashCode(), variant.rollCounter), size);
       variant.index = index == variant.index ? (index + 1) % size : index;
     }
     variant.clipName = roll.clipNames().get(variant.index);
